Lightweight fix for incorrect locale changes - DO NOT MERGE.
Previously, if the sim's MCC locale is reported AND
the user didn't have a persistent locale setting, then
the device locale would change.

This modifies the logic to only allow device locale changes
if the user has not explicitly set a locale and the user has
not completed setup.

An MCC-based update is still (erroneously) considered a
user-actioned change.

This is only intended to affect users that previously had an
unrecognized MCC. Android does a better job of recognizing
MCCs and we don't want those users that have received an OTA
to suddenly experience a locale change.

Bug: 19232829
Change-Id: I5b150573ec84f453bff54f3ca331b63e2c4d32d3