Fix for wifi issue on multisim
For release network request, request info is removed first
and then processed. Due to this, phone id cannot be retrieved
as there is no request info presnet, and default phone id of
sub1 is used. This is leading to issues in data call restore
after wifi is disabled.
Fix is to use dds phone id instead of default phone id.

Change-Id: I75bd02c42f9e1e97765d070f1821a53f7a096982
CRs-Fixed: 809766